+++ src/utils/.env
@@ -1,4 +1,4 @@
-PORT = 7000
+PORT = 3000
 
 # DB connection
 TYPEORM_CONNECTION = mysql
